fix errot fot telmask

+++ gammapy/spectrum/spectrum_extraction.py
@@ -303,8 +303,10 @@ class SpectrumExtraction(object):
 
         # Remove column TELMASK since raises an error when stacking it
         # see https://github.com/gammasky/hess-host-analyses/issues/34
+        
         for l in on_lists:
-            l.remove_column('TELMASK')
+            if(l.colnames=='TELMASK'):
+                l.remove_column('TELMASK')
 
         total_list = vstack(on_lists, join_type='inner', metadata_conflicts='silent')
         total_list.meta = None
@@ -318,7 +320,8 @@ class SpectrumExtraction(object):
         # Remove column TELMASK since raises an error when stacking it
         # see https://github.com/gammasky/hess-host-analyses/issues/34
         for l in off_lists:
-            l.remove_column('TELMASK')
+            if(l.colnames=='TELMASK'):
+                l.remove_column('TELMASK')
 
         total_list = vstack(off_lists, join_type='exact', metadata_conflicts='silent')
         total_list.meta = None
